+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Fixed
     - Inherited owner email address not displayed in project member list (#541)
     - App visibility check broken in ``project_detail.html`` (#546)
     - Invite accept for a category invoking Taskflow and causing a crash (#552)
+    - Project form ``parent`` forced to wrong value if user lacks role in parent category (#558)
 
 
 v0.8.1 (2020-04-24)